# app/public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/students": {
      "post": {
        "tags": [
          "Student API"
        ],
        "operationId": "App\\Controllers\\StudentController::create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "example": "junior@email.com"
                  },
                  "first_name": {
                    "example": "Bob"
                  },
                  "last_name": {
                    "example": "Jordan"
                  },
                  "phone": {
                    "example": "+7 495 1111111"
                  },
                  "skills": {
                    "example": "[2, 3, 4]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/students/{id}": {
      "get": {
        "tags": [
          "Student API"
        ],
        "operationId": "App\\Controllers\\StudentController::search",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Student API"
        ],
        "operationId": "App\\Controllers\\StudentController::delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Student API"
        ],
        "operationId": "App\\Controllers\\StudentController::update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "example": "junior@email.com"
                  },
                  "first_name": {
                    "example": "Bob"
                  },
                  "last_name": {
                    "example": "Jordan"
                  },
                  "phone": {
                    "example": "+7 495 3333333"
                  },
                  "enabled": {
                    "example": false
                  },
                  "skills": {
                    "example": "[3, 4, 5]"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/students/{id}/find-group": {
      "post": {
        "tags": [
          "Student API"
        ],
        "operationId": "App\\Controllers\\StudentController::findGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/teachers/{id}/find-group": {
      "post": {
        "tags": [
          "Teacher API"
        ],
        "operationId": "App\\Controllers\\TeacherController::findGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The identifier.",
            "required": true,
            "example": 1
          }
        ],
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    }
  }
}